# lpc-payment-backend-api/docs/swagger.yaml
openapi: 3.0.0
info:
  description: >
    This is the API for SDK for the Loyalty Payment Card.


    All dates are in ISO 8601 format (YYYY-MM-DD). They need to be converted
    from/to MMDDYYYY when interacting with SOAP.


    Pagination for GET requests is controlled via `page` and `per_page` query
    parameteres and presented via HTTP Headers:

    - `X-Total-Count` – total number of elements, eg. "X-Total-Count: 200". Response for GET requests in `payments` group don't contain this header.

    - `X-Page` – current page, eg. "X-Page: 1"

    - `X-Per-Page` – limit of elements per page, eg. "X-Per-Page: 20"

    - `Link` – link relations to first, last, next and previous pages, eg.
    "Link: \<https://example.com/api/v1/payments/cards?page=1&per_page=20\>;
    rel=\"first\",
    \<https://example.com/api/v1/payments/cards?page=2&per_page=20\>;
    rel=\"prev\",
    \<https://example.com/api/v1/payments/cards?page=4&per_page=20\>;
    rel=\"next\",
    \<https://example.com/api/v1/payments/cards?page=10&per_page=20\>;
    rel=\"last\""
  version: 1.0.0
  title: Poseidon LPC
security:
  - jwt: []
servers:
  - url: /api/v1
...
paths:
  /payments/customers:
    post:
      tags:
        - payments
      operationId: createCustomer
      summary: Create new customer in NKS
      description: |
        **Authorization**: JWT of any logged in user
        - InstitutionId must exist in CAS

        - BranchId must exist in CAS, if present

        - BirthDate must contain a valid date

        - CustomerType must contain valid value
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: DCSSCustomerCreate completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCreate'
        '400':
          description: >-
            Address/name field too long or some other
            fields are invald
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Institution or Branch doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '409':
          description: >-
            User already register as a customer in NKS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: >-
            System Error or Communication error between Webservice and
            underlying system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /payments/cards:
    get:
      tags:
        - payments
      operationId: listCards
      summary: Returns all card connected to the "input key".
      description: >
        **Authorization**: JWT of any logged in user


        This web service is PCI compliant if used correctly with the `pcimask`

        attribute. To enable PCI compliant behavior the optional `pcimask`

        header must be set to "true". The Source application needs to be

        approved for PCI compliant processing prior to using this value. When

        using the `pcimask` header with value "true" the service will adhere to

        PCI as follows:
          - `CardId` and `CardIdVNo` replaces `CardNo` and `VersionNo` as unique key towards a card record
          - `CardNo` and `VersionNo` may be ommited in requests
          - Wherever `CardNo` is needed as input, `CardId` shall be used instead
          - Wherever `VersionNo` not needed as input, `CardIdVNo` shall be used instead
          - `CardNo` will be masked according to PCI rules in responses

        If WS-consumer is running within a PCI compliant enviroment and
        `pcimask` is "false" or not present in the request:
          - `CardNo` and `VersionNo` shall be used as in previous versions of this service
          - `CardNo` and `VersionNo` is unique key towards a card record
          - `CardNo` will not be masked in responses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/PCIMaskParam'
        - $ref: '#/components/parameters/AccountNoParam'
        - $ref: '#/components/parameters/AccountOwnerParam'
        - $ref: '#/components/parameters/BranchParam'
        - $ref: '#/components/parameters/CardIdParam'
        - $ref: '#/components/parameters/CardIdVNoParam'
        - $ref: '#/components/parameters/CustomerIdParam'
        - $ref: '#/components/parameters/CustomerNoParam'
        - $ref: '#/components/parameters/ExtAgreementIdParam'
        - $ref: '#/components/parameters/ExternalProductIdParam'
        - $ref: '#/components/parameters/InstIdParam'
        - $ref: '#/components/parameters/OptionParam'
      responses:
        '200':
          description: DCSSCardList completed successfully
          headers:
            X-Page:
              $ref: '#/components/parameters/XPageParam'
            X-Per-Page:
              $ref: '#/components/parameters/XPerPageParam'
            Link:
              $ref: '#/components/parameters/LinkParam'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardList'
        '400':
          description: Invalid length of some parameteres
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: >-
            System Error or Communication error between Webservice and
            underlying system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
    post:
      tags:
        - payments
      operationId: createCard
      summary: Creates new card agreement in CAS
      description: >
        **Authorization**: JWT of any logged in user


        This web service is PCI compliant if used correctly with the `pcimask`
        attribute. To enable PCI compliant behavior the optional `pcimask`
        header must be set to "true". The Source application needs to be
        approved for PCI compliant processing prior to using this value. When
        using the `pcimask` header with value "true" the service will adhere to
        PCI as follows:
          - `CardId` and `CardIdVNo` replaces `CardNo` and `VersionNo` as unique key towards a card record
          - `CardNo` and `VersionNo` may be ommited in requests
          - Wherever `CardNo` is needed as input, `CardId` shall be used instead
          - Wherever `VersionNo` not needed as input, `CardIdVNo` shall be used instead
          - `CardNo` will be masked according to PCI rules in responses

        If WS-consumer is running within a PCI compliant enviroment and
        `pcimask` is "false" or not present in the request:
          - `CardNo` and `VersionNo` shall be used as in previous versions of this service
          - `CardNo` and `VersionNo` is unique key towards a card record
          - `CardNo` will not be masked in responses

        **Field validation:**
          - InstitutionId must exist in CAS
          - BranchId must be known in CAS, if present in txn
          - CustomerNo/Customer must be present in CAS
          - CustomerStatus must be "0" = OK, all other values will reject transaction
          - EmbossedName can not be emtpty string and max length according to product setup (default max length 26)
          - Expire must contain only numeric values and valid value
          - ProductId must exist and be valid in CAS
          - AccountNo must be within allowed accountrange
          - ShipmentMethodPin must be valid, valid values is set up on each product
          - ShipmentMethodCard must be valid, valid values is set up on each product
          - CreditCard, credit account must exist in CAS.
          - CredLimit must match product setup, between max and min values
          - CredLimit must be below the accounts creditlimit
          - BulkId must exist in CAS
      parameters:
        - $ref: '#/components/parameters/PCIMaskParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreate'
      responses:
        '200':
          description: DCSSCardCreate completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardCreateResponse'
        '400':
          description: Embossedname to long or some other field validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 'Institution, Branch, Consumer, Product or Bulk doesn''t exist'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: >-
            System Error or Communication error between Webservice and
            underlying system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /payments/history:
    get:
      tags:
        - payments
      operationId: cardHistory
      summary: Card history
      description: >
        **Authorization**: JWT of any logged in user


        Returns list of changes to the card during its lifecycle. Client apps.
        must use one, and only one, of the input key-fields.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/AccountNoParam'
        - $ref: '#/components/parameters/CardNoParam'
        - $ref: '#/components/parameters/CustomerIdStringParam'
        - $ref: '#/components/parameters/InstIdParam'
        - $ref: '#/components/parameters/VersionNoParam'
      responses:
        '200':
          description: DCSSCardHistory completed successfully
          headers:
            X-Page:
              $ref: '#/components/parameters/XPageParam'
            X-Per-Page:
              $ref: '#/components/parameters/XPerPageParam'
            Link:
              $ref: '#/components/parameters/LinkParam'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardHistory'
        '400':
          description: 'No key-field or more than one, or invalid'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: >-
            System Error or Communication error between Webservice and
            underlying system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
...
components:
  parameters:
    PCIMaskParam:
      name: pcimask
      in: header
      description: Whether to enable PCI compliant behavior
      example: true
      schema:
        type: string
        default: false
...
    XPageParam:
      name: X-Page
      in: header
      example: 1
      description: Current page
      schema:
        type: integer
    XPerPageParam:
      name: X-Per-Page
      in: header
      example: 20
      description: Total number of pages
      schema:
        type: integer
    LinkParam:
      name: Link
      in: header
      example: >-
        Link: <https://example.com/api/v1/payments/cards?page=1&per_page=20>;
        rel="first",
        <https://example.com/api/v1/payments/cards?page=2&per_page=20>;
        rel="next",
        <https://example.com/api/v1/payments/cards?page=10&per_page=20>;
        rel="last"
      description: relations to other pages
      schema:
        type: string
    PageParam:
      name: page
      in: query
      example: 1
      schema:
        type: integer
        default: 1
    PerPageParam:
      name: per_page
      in: query
      example: 20
      schema:
        type: integer
        default: 20
    AccountNoParam:
      name: AccountNo
      in: query
      description: Account Number
      schema:
        minLength: 1
        type: string
        maxLength: 35
        example: XXXXXXXXXXXXXXXXXXXXXXXX
    AccountOwnerParam:
      name: AccountOwner
      in: query
      description: Account owners customer number
      schema:
        type: string
        maxLength: 35
        example: XXXXXXXXXX
    BranchParam:
      name: Branch
      in: query
      description: Branch ID. Identifies which branch the card belongs too
      schema:
        type: string
        minLength: 3
        maxLength: 9
        example: XXXXXXXXX
    CardIdParam:
      name: CardId
      in: query
      description: CardIdent
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 10000000000000000
        example: 1234
    CardNoParam:
      name: CardNo
      in: query
      description: Card Number
      schema:
        type: string
        minLength: 11
        maxLength: 19
        example: 1234123412341234
    CustomerIdParam:
      name: CustomerId
      in: query
      description: >-
        CustomerIdentification. Customer identification in external system. This
        id is used to match the card with customerdata received from external
        system. This might be a sequnce number assigned, or any other numbering
        system used in the external system
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 100000000000000000
        example: 1234
    CustomerIdStringParam:
      name: CustomerId
      in: query
      description: CustomerId
      schema:
        type: string
        minLength: 11
        maxLength: 17
        example: 1234
    CardIdVNoParam:
      name: CardIdVNo
      in: query
      description: CardId Version Number
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 999
        example: 1
    CustomerNoParam:
      name: CustomerNo
      in: query
      description: >-
        CustomerNumber. CustomerNo (Person/Foretaksnr) Typically contains
        "Social security number" or equivalent data for this customer. Must be
        present in CAS
      schema:
        type: string
        minLength: 1
        maxLength: 35
        example: XXXXXXXXXXXX
    ExtAgreementIdParam:
      name: ExtAgreementId
      in: query
      description: Agreement id external to CAS
      schema:
        type: string
        maxLength: 19
        example: XXXXXXXXXXXXXXXXXXX
    ExternalProductIdParam:
      name: ExternalProductId
      in: query
      description: >-
        Optional input external product identifier. If present in request only
        cards eligible for this external product will be present in respons
        message. If the card has an agreement for this external productId then
        the agreementNo item will be present in response messages
      schema:
        type: string
        minLength: 1
        maxLength: 20
        example: XXXXXXXXXX
    InstIdParam:
      name: InstId
      in: query
      description: >-
        Institution ID. Identifies which bank that has issued the card. Must
        exist in CAS
      schema:
        type: string
        minLength: 4
        maxLength: 9
        example: XXXXXXXXX
    OptionParam:
      name: Option
      in: query
      description: |
        Search options.
        **CAS**
          - 0 = Open and Blocked cards are listed
          - 1 = only open card are listed
          - 2 only blocked cards are listed
          - 3 = only deleted cards are listed
          - 9 All cards is listed regardles of status
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 9
        default: null
        example: 0
    VersionNoParam:
      name: VersionNo
      in: query
      description: Card Version Number
      schema:
        type: string
        minLength: 1
        maxLength: 3
        example: '1'
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CardCreateResponse:
      type: object
      properties:
        CardId:
          type: integer
          format: int64
          minimum: 1
          maximum: 10000000000000000
          description: CardIdent
          example: 1234
        CardIdVNo:
          type: integer
          format: int32
          minimum: 1
          maximum: 999
          description: CardId Version Number
          example: 1
        CardNo:
          type: string
          minLength: 11
          maxLength: 19
          description: >-
            Card Number. Field is masked according to PCI if header
            `pcimask`=truer, otherwise CardNo will be returned unmasked
          example: XXXXXXXXXXXXXXXX
        TxAccountNo:
          type: integer
          format: int64
          minimum: 1
          maximum: 4999999999999999
          description: >-
            Transaction Account number (Card Number Alias). Used for
            CreditCards.
          example: 1234
        VersionNo:
          type: string
          minLength: 1
          maxLength: 3
          description: Version number generated by CAS
          example: '1'
...
    CustomerCreate:
      type: object
      required:
        - CustomerNo
        - InstId
      properties:
        CustomerNo:
          type: string
          minLength: 1
          maxLength: 35
          writeOnly: true
          description: CustomerNumber
          example: XXXXXXXXXXXXXXXXXXXXXXXXX
        InstId:
          type: string
          minLength: 4
          maxLength: 9
          writeOnly: true
          description: Institution ID. Identifies which bank that has issued the card.
          example: XXXXXXXXX
        AccountList:
          type: string
          maxLength: 3600
          writeOnly: true
          description: >
            The account numbers to be used when creating debit cards must be
            present in AccountList. Each account is a TLV structure
              - Mandatory tags:
                - TAG "01" = AccountNo,
                - TAG "02" = PaymentScheme Values: D debit, C credit, S shadowbalance
              - Optional tags:
                - TAG "06" = ProductCode (LF, as received from LF)
                - TAG "07" = Accountowners CustomerNo
                - TAG "08" = ProductName (LF, as received from LF)
                - TAG "09" = Relation-code (LF-only)
                - TAG "10" = Clearing code (LF-Only)
              - Optional tags, for future use:
                - TAG "03" = Currencycode
                - TAG "04" = ProductList, identifies which cards may be issued on this account
                - TAG "05" = MakeChecksts of the account, if unchecked cards should not be issued. Default value is "0", MakeChecksts not used/relevant for this account
        Address1:
          type: string
          maxLength: 40
          writeOnly: true
          description: Address line 1.
        Address2:
          type: string
          maxLength: 40
          writeOnly: true
          description: Address line 2.
        Address3:
          type: string
          maxLength: 40
          writeOnly: true
          description: Address line 3.
        Address4:
          type: string
          maxLength: 40
          writeOnly: true
          description: Address line 4.
        Address5:
          type: string
          maxLength: 40
          writeOnly: true
          description: Address line 5.
        Address6:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 6.
        AltAddress1:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 1.
        AltAddress2:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 2.
        AltAddress3:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 3.
        AltAddress4:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 4.
        AltAddress5:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 5.
        AltAddress6:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate address line 6.
        AltCity:
          type: string
          maxLength: 40
          writeOnly: true
          description: AlternateCity
          example: Warszawa
        AltCountry:
          type: string
          maxLength: 40
          writeOnly: true
          description: Alternate country
          example: Poland
        AltPostalCode:
          type: string
          maxLength: 9
          writeOnly: true
          description: Alternate postal code
          example: 01-001
        AuthA1:
          type: string
          maxLength: 64
          writeOnly: true
          description: Answer to question No 1
        AuthA2:
          type: string
          maxLength: 64
          writeOnly: true
          description: Answer to question No 1
        AuthA3:
          type: string
          maxLength: 64
          writeOnly: true
          description: Answer to question No 1
        AuthQ1:
          type: string
          maxLength: 124
          writeOnly: true
          description: Question No 1 used for customer authentication
        AuthQ2:
          type: string
          maxLength: 124
          writeOnly: true
          description: Question No 2 used for customer authentication
        AuthQ3:
          type: string
          maxLength: 124
          writeOnly: true
          description: Question No 3 used for customer authentication
        BirthDate:
          type: string
          format: date
          writeOnly: true
          description: Date of birth for customer
          example: '1990-01-01'
        Branch:
          type: string
          minLength: 2
          maxLength: 9
          writeOnly: true
          description: 'Branch ID, must exist in CAS if present'
          example: XXXXXXXXX
          default: InstId
        BusinessPhone:
          type: string
          maxLength: 16
          writeOnly: true
          description: Business phone
          example: '+4822111111111'
        Citizenship:
          type: string
          maxLength: 25
          writeOnly: true
          description: Citizenship customer
        City:
          type: string
          maxLength: 30
          writeOnly: true
          description: Postal city
          example: Warsaw
        CompanyName:
          type: string
          maxLength: 30
          writeOnly: true
          description: Company name. Only used for `customerType` = 6
        Country:
          type: string
          maxLength: 30
          writeOnly: true
          description: Country
          example: Poland
        CountryCode:
          type: string
          maxLength: 3
          writeOnly: true
          description: Country code
          example: PL
        CreditAllowed:
          type: string
          maxLength: 1
          writeOnly: true
          description: Flag saying if customer is allowed for creditcard
          example: 'Y'
        CreditScodeAmount:
          type: number
          format: float
          writeOnly: true
          description: Creditscore amount
          example: 1000.42
        CustomerCategory:
          type: string
          maxLength: 20
          writeOnly: true
          description: Customer category. This is a freetext field.
        CustomerId:
          type: integer
          format: int64
          minimum: 1
          maximum: 100000000000000000
          readOnly: true
          description: CustomerIdent
          example: 1234
        CustomerInfo:
          type: string
          maxLength: 50
          writeOnly: true
          description: Customer category. This is a freetext field.
        CustomerStatus:
          type: string
          maxLength: 4
          writeOnly: true
          description: |
            Valid values:
            - 0 - Customer OK
            - 1 - Customer deceased
            - 2 - Customer bankrupt
            - 3 - Customer closed
          example: '0'
        CustomerType:
          type: string
          maxLength: 4
          writeOnly: true
          description: >
            Legal values:

            - 1 - Customer is a person. CustomerNo is a social security number

            - 2 - Customer is an organisation. Card creation is not allowed.
            CustomerNo is an organisation number

            - 3 - Customer is a person with D number. CustomerNo is a D number

            - 5 - Customer is an organisation. Card creation allowed. CustomerNo
            is an organisation number

            - 6 - Business customer. Card creation NOT allowed.
        DigiPassSerialNo:
          type: string
          maxLength: 12
          writeOnly: true
          description: DigiPass serial number
          example: XXXXXXXXXXXX
        DrvLicNo:
          type: string
          maxLength: 24
          writeOnly: true
          description: Driver licence number
        EmailHome:
          type: string
          maxLength: 75
          writeOnly: true
          description: Email home
        EmailWork:
          type: string
          maxLength: 75
          writeOnly: true
          description: Email work
        EmbossedName:
          type: string
          maxLength: 26
          writeOnly: true
          description: As received from external cust-sys
        EmbossedNameLong:
          type: string
          maxLength: 30
          writeOnly: true
          description: Embossedname long variant. As received from external cust-sys
        Felt1:
          type: string
          maxLength: 20
          writeOnly: true
          description: Kunde Felt1 Felt1 create
        FirstName:
          type: string
          maxLength: 40
          writeOnly: true
          description: >-
            Field is used in card shipment on envelope, field is also encoded in
            magnetic stripe1
          example: John
        GeoCode:
          type: string
          maxLength: 4
          writeOnly: true
          description: >-
            Customer geografic code. if norwegian adress then "kommunenr".
            Otherwise then 9 + country code ISOnumeric
          example: '9616'
        IndustryCode:
          type: integer
          format: int32
          minimum: 0
          maximum: 99999
          writeOnly: true
          description: Accountowners Industry code
        Language:
          type: string
          minLength: 2
          maxLength: 2
          writeOnly: true
          description: Customers language code ISO-639-1
          example: pl
        LastName:
          type: string
          maxLength: 40
          writeOnly: true
          description: Last Name
          example: Doe
        LegalCustomer:
          type: boolean
          writeOnly: true
          description: Flag saying if legal customer
          example: true
        MakeCheckSts:
          type: integer
          format: int32
          minimum: 1
          maximum: 99
          nullable: true
          writeOnly: true
          default: null
          description: Status for maker checker
          example: 1
        Memo:
          type: string
          maxLength: 2000
          writeOnly: true
          description: Customer notes
        MobilePhone:
          type: string
          maxLength: 20
          writeOnly: true
          description: Mobile phone
          example: '+48602602602'
        PassportNo:
          type: string
          maxLength: 24
          writeOnly: true
          description: Passport number
          example: XXXXXXXXXX
        PostalCode:
          type: string
          maxLength: 9
          writeOnly: true
          description: Postal code
          example: 01-001
        RelManagerEmail:
          type: string
          maxLength: 75
          writeOnly: true
          description: Relationship manager email
          example: manager@example.com
        RelManagerName:
          type: string
          maxLength: 60
          writeOnly: true
          description: Relationship manager name
          example: John Doe
        RelManagerTlf:
          type: string
          maxLength: 20
          writeOnly: true
          description: Relationship manager phone number
          example: '+48602602602'
        ScoreDate:
          type: string
          format: date
          writeOnly: true
          description: Score date
          example: '2019-01-01'
        SectorCode:
          type: integer
          format: int32
          minimum: 1
          maximum: 99999
          writeOnly: true
          description: Accountowners SectorCode
          example: 1
        SignatureCardDate:
          type: string
          format: date
          writeOnly: true
          description: Date when SignatureCard was created
          example: '2019-01-01'
        SignaturRef:
          type: integer
          format: int32
          minimum: 1
          maximum: 99999999
          writeOnly: true
          description: Signatureref
          example: 1234
        TDSecure:
          type: string
          maxLength: 20
          writeOnly: true
          description: 3D Secure password
        Title:
          type: string
          maxLength: 8
          writeOnly: true
          description: CustomerTitle
          example: Mr.
        Tlf:
          type: string
          maxLength: 16
          writeOnly: true
          description: Phone number
          example: '+48602602602'
        VIPsts:
          type: string
          maxLength: 8
          writeOnly: true
          description: VIP status
...
    BadRequest:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          description: Error message
          example: Bad request
    Unauthorized:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          description: Error message
          example: Unauthorized
...
    NotFound:
      type: object
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          description: Error message
          example: Not found
    Conflict:
      type: object
      properties:
        code:
          type: integer
          example: 409
        message:
          type: string
          description: Error message
          example: Conflict
    ServerError:
      type: object
      properties:
        code:
          type: integer
          example: 500
        message:
          type: string
          description: Error message
          example: Internal server error